refactor(graph): move goals prompt-rendering out of types.rs
goals/types.rs mixed the ThreadGoal/ThreadGoalStatus/GoalProgress data
model with active_goal_context_block, a prompt-rendering function —
presentation, not state. Moved to prompt.rs. No behavior change.
