+++ src/graph/goals/mod.rs
@@ -13,15 +13,15 @@
 //! primitive is provider-neutral and drives off the graph runtime.
 
 mod continuation;
+mod prompt;
 pub mod store;
 mod tool;
 mod types;
 
 pub use continuation::{goal_gate_node, note_user_turn, run_continuation_tick};
+pub use prompt::active_goal_context_block;
 pub use tool::{GoalTool, GoalToolKind, goal_tools, register_goal_tools};
-pub use types::{
-    GoalProgress, ThreadGoal, ThreadGoalStatus, TurnOutcome, active_goal_context_block,
-};
+pub use types::{GoalProgress, ThreadGoal, ThreadGoalStatus, TurnOutcome};
 
 #[cfg(test)]
 mod test;
